+++ skeleton/ForceRegistry.h
@@ -6,7 +6,7 @@ typedef std::pair<ForceGenerator*, Particle*> forceParticlePair;
 
 class ForceRegistry : public std::multimap<ForceGenerator*, Particle*> {
 public:
-	void updateForces(double duration);
+	void updateForces(float duration);
 	void addRegistry(Particle*p, ForceGenerator* f);
 	void deleteParticle(Particle*p);
 	void deleteForce(ForceGenerator*f);
